Remove explicit modification of privacy indicators flag
Android 16 introduces a new restriction that only allowlisted
DeviceConfig flags can be modified by the shell user. The
camera_mic_icons_enabled flag is not on the allowlist since it
was only intended to be used during the initial feature development
and should be enabled by default. This commit updates the test
that currently modifies this flag to only verify that the flag
is enabled.

Bug: 390242290
Flag: android.security.protect_device_config_flags
Relnote: N/A
Test: atest CtsPermissionUiTestCases:CameraMicIndicatorsPermissionTest
Change-Id: I11ed9c3b6c29a32557add30c5894e728ca9712f7
LOW_COVERAGE_REASON=TEST_ONLY