WearAppPermissionGroupsHelper: Prioritize platform permissions on AppInfo page.
The HEALTH permission group is created by the healthfitness mainline module and not the OS_PKG. This meant that even though it is a platform permission group that it was getting sorted to the bottom of the list. Now we partition permissions based on whether they are platform permissions to ensure it takes it's rightful spot near the top.

Bug: 404941005
Test: Locally built, verified ordering was fixed.
Flag: EXEMPT bugfix
Relnote: Fixing permission order on Wear App info page.
LOW_COVERAGE_REASON=OTHER_FORM_FACTOR
Change-Id: I33d901d330cd963494f83a0e4b778db3aebfb529