Permissions: Fix admin restricted CTS Test.
The PermissionUsageViewModelTest in the test suite was adding a health permission to the platform list which was causing the HEALTH permission group to be considered restricted. When we checked with the group it would then always return restricted even before we added the permission we were checking to the list.

Now we remove the explicit group from being passed in. The point of this test is mostly to ensure that permissions that are added to the list become restricted rather than being about health specifically. The inprocess tests cover the health permissions getting added.

Bug: 403921690
Test: atest PermissionControllerMockingTests
Flag: EXEMPT bugfix
Relnote: N/A
LOW_COVERAGE_REASON=TEST_ONLY
Change-Id: Ibba609dbe28c637206f17f375e273888d565f5c9